Machine Learning Models
The process of implementing and evaluating the machine learning models described in the code focuses on the classification of emotional states (valence and arousal) using pre-processed and normalised physiological signal data from the DREAMER dataset. This code includes the key steps of preprocessing, model building and evaluation, as well as saving the best trained models for later use.